+++ src/main/java/com/telepathicgrunt/blame/mixin/WorldSettingsImportInnerClassMixin.java
@@ -1,26 +1,20 @@
 package com.telepathicgrunt.blame.mixin;
 
 import com.google.gson.JsonElement;
-import com.mojang.serialization.Codec;
 import com.mojang.serialization.DataResult;
 import com.mojang.serialization.Encoder;
 import com.mojang.serialization.Lifecycle;
 import com.telepathicgrunt.blame.main.WorldSettingsImportBlame;
 import net.minecraft.util.RegistryKey;
-import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.registry.DynamicRegistries;
-import net.minecraft.util.registry.MutableRegistry;
-import net.minecraft.util.registry.Registry;
 import net.minecraft.util.registry.WorldSettingsImport;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
 import java.util.Optional;
-import java.util.function.Supplier;
 
 /* @author - TelepathicGrunt
  *